+++ spharpy/samplings/coordinates.py
@@ -8,7 +8,8 @@ class SamplingSphere(pf.Coordinates):
 
     def __init__(
             self, x=None, y=None, z=None, n_max=None, weights: np.array = None,
-            comment: str = ""):
+            quadrature: bool = False, comment: str = "",
+            radius_tolerance=1e-6):
         r"""
         Create a SamplingSphere class object from a set of points on a sphere.
 
@@ -25,27 +26,43 @@ class SamplingSphere(pf.Coordinates):
         z : ndarray, number
             Z coordinate of a right handed Cartesian coordinate system in
             meters (-\infty < z < \infty).
+        n_max : int, optional
+            Maximum spherical harmonic order of the sampling grid.
+            The default is ``None``.
         weights: array like, number, optional
             Weighting factors for coordinate points. Their sum must equal to
             the integral over the unit sphere, which is :math:`4\pi`.
             The `shape` of the array must match the `shape` of the individual
             coordinate arrays. The default is ``None``, which means that no
             weights are used.
+        quadrature : bool, optional
+            Flag that indicates if points belong to a quadrature, which
+            requires that `weights` is not ``None``. The default is ``False``.
         comment : str, optional
             Comment about the stored coordinate points. The default is
             ``""``, which initializes an empty string.
-        sh_order : int, optional
-            Maximum spherical harmonic order of the sampling grid.
-            The default is ``None``.
+        radius_tolerance : float, optional
+            All points that are stored in a SamplingSphere must have the same
+            radius and an error is raised if the maximum deviation from the
+            mean radius exceeds this tolerance. The default of ``1e-6`` meter
+            is intended to allow for some numerical inaccuracy.
         """
+        self._radius_tolerance = None
+        self.radius_tolerance = radius_tolerance
+
         pf.Coordinates.__init__(
             self, x, y, z, weights=weights, comment=comment)
         self._n_max = n_max
 
+        # initialize and set quadrature
+        self._quadrature = None
+        self.quadrature = quadrature
+
     @classmethod
     def from_cartesian(
             cls, x, y, z, n_max=None, weights: np.array = None,
-            comment: str = ""):
+            quadrature: bool = False, comment: str = "",
+            radius_tolerance: float = 1e-6):
         r"""
         Create a Coordinates class object from a set of points on a sphere.
 
@@ -71,9 +88,17 @@ class SamplingSphere(pf.Coordinates):
             The `shape` of the array must match the `shape` of the individual
             coordinate arrays. The default is ``None``, which means that no
             weights are used.
+        quadrature : bool, optional
+            Flag that indicates if points belong to a quadrature, which
+            requires that `weights` is not ``None``. The default is ``False``.
         comment : str, optional
             Comment about the stored coordinate points. The default is
             ``""``, which initializes an empty string.
+        radius_tolerance : float, optional
+            All points that are stored in a SamplingSphere must have the same
+            radius and an error is raised if the maximum deviation from the
+            mean radius exceeds this tolerance. The default of ``1e-6`` meter
+            is intended to allow for some numerical inaccuracy.
 
         Examples
         --------
@@ -85,12 +110,14 @@ class SamplingSphere(pf.Coordinates):
         >>> sampling = pf.SamplingSphere(0, 0, 1)
         """
         return cls(
-            x, y, z, weights=weights, comment=comment, n_max=n_max)
+            x, y, z, weights=weights, comment=comment, n_max=n_max,
+            quadrature=quadrature, radius_tolerance=radius_tolerance)
 
     @classmethod
     def from_spherical_elevation(
             cls, azimuth, elevation, radius, n_max=None,
-            weights: np.array = None, comment: str = ""):
+            weights: np.array = None, quadrature: bool = False,
+            comment: str = "", radius_tolerance: float = 1e-6):
         """Create a Coordinates class object from a set of points on a sphere.
 
         See :py:mod:`pyfar.classes.coordinates` for  more information.
@@ -116,9 +143,17 @@ class SamplingSphere(pf.Coordinates):
             The `shape` of the array must match the `shape` of the individual
             coordinate arrays. The default is ``None``, which means that no
             weights are used.
+        quadrature : bool, optional
+            Flag that indicates if points belong to a quadrature, which
+            requires that `weights` is not ``None``. The default is ``False``.
         comment : str, optional
             Comment about the stored coordinate points. The default is
             ``""``, which initializes an empty string.
+        radius_tolerance : float, optional
+            All points that are stored in a SamplingSphere must have the same
+            radius and an error is raised if the maximum deviation from the
+            mean radius exceeds this tolerance. The default of ``1e-6`` meter
+            is intended to allow for some numerical inaccuracy.
 
         Examples
         --------
@@ -127,15 +162,18 @@ class SamplingSphere(pf.Coordinates):
         >>> sampling = pf.SamplingSphere.from_spherical_elevation(0, 0, 1)
         """
 
-        x, y, z = sph2cart(azimuth, np.pi / 2 - elevation, radius)
+        x, y, z = sph2cart(
+            azimuth, np.pi / 2 - np.atleast_1d(elevation), radius)
         return cls(
-            x, y, z, weights=weights, comment=comment, n_max=n_max)
+            x, y, z, weights=weights, comment=comment, n_max=n_max,
+            quadrature=quadrature, radius_tolerance=radius_tolerance)
 
     @classmethod
     def from_spherical_colatitude(
             cls, azimuth, colatitude, radius, n_max=None,
-            weights: np.array = None, comment: str = ""):
-        """Create a Coordinates class object from a set of points on a sphere.
+            weights: np.array = None, quadrature: bool = False,
+            comment: str = "", radius_tolerance: float = 1e-6):
+        r"""Create a Coordinates class object from a set of points on a sphere.
 
         See :py:mod:`pyfar.classes.coordinates` for  more information.
 
@@ -160,9 +198,17 @@ class SamplingSphere(pf.Coordinates):
             The `shape` of the array must match the `shape` of the individual
             coordinate arrays. The default is ``None``, which means that no
             weights are used.
+        quadrature : bool, optional
+            Flag that indicates if points belong to a quadrature, which
+            requires that `weights` is not ``None``. The default is ``False``.
         comment : str, optional
             Comment about the stored coordinate points. The default is
             ``""``, which initializes an empty string.
+        radius_tolerance : float, optional
+            All points that are stored in a SamplingSphere must have the same
+            radius and an error is raised if the maximum deviation from the
+            mean radius exceeds this tolerance. The default of ``1e-6`` meter
+            is intended to allow for some numerical inaccuracy.
 
         Examples
         --------
@@ -173,13 +219,15 @@ class SamplingSphere(pf.Coordinates):
 
         x, y, z = sph2cart(azimuth, colatitude, radius)
         return cls(
-            x, y, z, weights=weights, comment=comment, n_max=n_max)
+            x, y, z, weights=weights, comment=comment, n_max=n_max,
+            quadrature=quadrature, radius_tolerance=radius_tolerance)
 
     @classmethod
     def from_spherical_side(
             cls, lateral, polar, radius, n_max=None,
-            weights: np.array = None, comment: str = ""):
-        """Create a Coordinates class object from a set of points on a sphere.
+            weights: np.array = None, quadrature: bool = False,
+            comment: str = "", radius_tolerance: float = 1e-6):
+        r"""Create a Coordinates class object from a set of points on a sphere.
 
         See :py:mod:`pyfar.classes.coordinates` for  more information.
 
@@ -203,9 +251,17 @@ class SamplingSphere(pf.Coordinates):
             The `shape` of the array must match the `shape` of the individual
             coordinate arrays. The default is ``None``, which means that no
             weights are used.
+        quadrature : bool, optional
+            Flag that indicates if points belong to a quadrature, which
+            requires that `weights` is not ``None``. The default is ``False``.
         comment : str, optional
             Comment about the stored coordinate points. The default is
             ``""``, which initializes an empty string.
+        radius_tolerance : float, optional
+            All points that are stored in a SamplingSphere must have the same
+            radius and an error is raised if the maximum deviation from the
+            mean radius exceeds this tolerance. The default of ``1e-6`` meter
+            is intended to allow for some numerical inaccuracy.
 
         Examples
         --------
@@ -214,15 +270,18 @@ class SamplingSphere(pf.Coordinates):
         >>> sampling = pf.SamplingSphere.from_spherical_side(0, 0, 1)
         """
 
-        x, z, y = sph2cart(polar, np.pi / 2 - lateral, radius)
+        x, z, y = sph2cart(
+            polar, np.pi / 2 - np.atleast_1d(lateral), radius)
         return cls(
-            x, y, z, weights=weights, comment=comment, n_max=n_max)
+            x, y, z, weights=weights, comment=comment, n_max=n_max,
+            quadrature=quadrature, radius_tolerance=radius_tolerance)
 
     @classmethod
     def from_spherical_front(
             cls, frontal, upper, radius, n_max=None, weights: np.array = None,
-            comment: str = ""):
-        """Create a Coordinates class object from a set of points on a sphere.
+            quadrature: bool = False, comment: str = "",
+            radius_tolerance: float = 1e-6):
+        r"""Create a Coordinates class object from a set of points on a sphere.
 
         See :py:mod:`pyfar.classes.coordinates` for  more information.
 
@@ -246,9 +305,17 @@ class SamplingSphere(pf.Coordinates):
             The `shape` of the array must match the `shape` of the individual
             coordinate arrays. The default is ``None``, which means that no
             weights are used.
+        quadrature : bool, optional
+            Flag that indicates if points belong to a quadrature, which
+            requires that `weights` is not ``None``. The default is ``False``.
         comment : str, optional
             Comment about the stored coordinate points. The default is
             ``""``, which initializes an empty string.
+        radius_tolerance : float, optional
+            All points that are stored in a SamplingSphere must have the same
+            radius and an error is raised if the maximum deviation from the
+            mean radius exceeds this tolerance. The default of ``1e-6`` meter
+            is intended to allow for some numerical inaccuracy.
 
         Examples
         --------
@@ -259,13 +326,15 @@ class SamplingSphere(pf.Coordinates):
 
         y, z, x = sph2cart(frontal, upper, radius)
         return cls(
-            x, y, z, weights=weights, comment=comment, n_max=n_max)
+            x, y, z, weights=weights, comment=comment, n_max=n_max,
+            quadrature=quadrature, radius_tolerance=radius_tolerance)
 
     @classmethod
     def from_cylindrical(
             cls, azimuth, z, rho, n_max=None, weights: np.array = None,
-            comment: str = ""):
-        """Create a Coordinates class object from a set of points on a sphere.
+            quadrature: bool = False, comment: str = "",
+            radius_tolerance: float = 1e-6):
+        r"""Create a Coordinates class object from a set of points on a sphere.
 
         See :py:mod:`pyfar.classes.coordinates` for  more information.
 
@@ -289,9 +358,17 @@ class SamplingSphere(pf.Coordinates):
             The `shape` of the array must match the `shape` of the individual
             coordinate arrays. The default is ``None``, which means that no
             weights are used.
+        quadrature : bool, optional
+            Flag that indicates if points belong to a quadrature, which
+            requires that `weights` is not ``None``. The default is ``False``.
         comment : str, optional
             Comment about the stored coordinate points. The default is
             ``""``, which initializes an empty string.
+        radius_tolerance : float, optional
+            All points that are stored in a SamplingSphere must have the same
+            radius and an error is raised if the maximum deviation from the
+            mean radius exceeds this tolerance. The default of ``1e-6`` meter
+            is intended to allow for some numerical inaccuracy.
 
         Examples
         --------
@@ -302,7 +379,8 @@ class SamplingSphere(pf.Coordinates):
 
         x, y, z = cyl2cart(azimuth, z, rho)
         return cls(
-            x, y, z, weights=weights, comment=comment, n_max=n_max)
+            x, y, z, weights=weights, comment=comment, n_max=n_max,
+            quadrature=quadrature, radius_tolerance=radius_tolerance)
 
     @property
     def n_max(self):
@@ -317,6 +395,52 @@ class SamplingSphere(pf.Coordinates):
             self._n_max = None
         else:
             self._n_max = int(value)
+
+    @property
+    def radius_tolerance(self):
+        """Get or set the radius tolerance in meter."""
+        return self._radius_tolerance
+
+    @radius_tolerance.setter
+    def radius_tolerance(self, value):
+        """Get or set the radius tolerance in meter."""
+
+        # check input
+        if not isinstance(value, (int, float)) or value <= 0:
+            raise ValueError(
+                'The radius tolerance must be a number greater than zero')
+
+        current_tolerance = self.radius_tolerance
+        self._radius_tolerance = float(value)
+
+        # Check if points meet new tolerance if points exist
+        if hasattr(self, 'x'):
+            try:
+                self._check_points(self._x, self._y, self._z)
+            except ValueError as e:
+                # revert setting the tolerance and raise the error
+                self._radius_tolerance = current_tolerance
+                raise e
+
+    def _check_points(self, x, y, z):
+        """Check input data before setting coordinates"""
+
+        # convert to numpy arrays of the same shape
+        x, y, z = super()._check_points(x, y, z)
+
+        # check for equal radius
+        radius = np.sqrt(x.flatten()**2 + y.flatten()**2 + z.flatten()**2)
+        radius_delta = np.max(radius) - np.min(radius)
+        if radius_delta > self.radius_tolerance:
+            raise ValueError(
+                'All points must have the same radius but the difference '
+                f'between the minimum and maximum radius is {radius_delta:.3g}'
+                ' m, which exceeds the tolerance of '
+                f'{self.radius_tolerance:.3g} m. The tolerance can be changed '
+                'using SamplingSphere.radius_tolerance.')
+
+        return x, y, z
+
 
     def _check_weights(self, weights):
         r"""Check if the weights are valid.
@@ -361,3 +485,25 @@ class SamplingSphere(pf.Coordinates):
         Their sum must equal to :math:`4\pi`.
         """
         super(__class__, type(self)).weights.fset(self, weights)
+
+    @property
+    def quadrature(self):
+        """Get or set the quadrature flag."""
+        return self._quadrature
+
+    @quadrature.setter
+    def quadrature(self, value):
+        """Get or set the quadrature flag."""
+
+        # check input
+        if not isinstance(value, bool):
+            raise TypeError(
+                f'quadrature must be True or False but is {value}')
+
+        # check if weights are set
+        # (if they are the weights setter enforces that they sum to 4 pi)
+        if self.weights is None and value:
+            raise ValueError(
+                'quadrature can not be True because the weights are None')
+
+        self._quadrature = value